% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_rings.R
\name{add_ring_group_by_number}
\alias{add_ring_group_by_number}
\title{Title}
\usage{
add_ring_group_by_number(.x, ..., ring, group_variable = grp)
}
\arguments{
\item{group_variable}{}
}
\description{
Title
}
