% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_gis_join_areas.R
\name{admin_areas_join}
\alias{admin_areas_join}
\title{Title}
\usage{
admin_areas_join(
  .x,
  lan_sf = lan,
  kommun_sf = kommun,
  distrikt_sf = distrikt,
  rapportomrade_sf = rapportomrade,
  id = FALSE
)
}
\arguments{
\item{id}{}
}
\description{
Title
}
