% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_gis_join_areas.R
\name{admin_areas_load}
\alias{admin_areas_load}
\title{Title}
\usage{
admin_areas_load(
  db = "G:/Naturvard/Naturvard.gpkg",
  layer = c("admin_kommun_fastk", "admin_lan_fastk", "admin_distrikt",
    "admin_landskap_fran_distrikt", "admin_landskap_artportalen",
    "admin_rapportomrade_artportalen"),
  import_to_sf = c("kommun", "lan", "distrikt", "landskap", "landskap_ap",
    "rapportomrade")
)
}
\arguments{
\item{import_to_sf}{}
}
\description{
Title
}
