% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_indicator.R
\name{brood_size_boot}
\alias{brood_size_boot}
\title{Title}
\usage{
brood_size_boot(
  x = c(24, 58, 9),
  n.boot = 10000,
  q = c(0.025, 0.5, 0.975),
  min = 15,
  size = sum(x),
  replace = TRUE,
  plot = TRUE
)
}
\arguments{
\item{plot}{}
}
\description{
Title
}
