% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_mean_geometric.R
\name{ci}
\alias{ci}
\title{Title}
\usage{
ci(
  x,
  type = c("geometric", "arithmetic", "harmonic"),
  dist = c("t", "norm"),
  alpha = 0.05,
  range = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{na.rm}{}
}
\description{
Title
}
