% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helper_functions.R
\name{convert_to_logical}
\alias{convert_to_logical}
\title{Title}
\usage{
convert_to_logical(.x, cols, yes_value = "Ja", no_value = "Nej")
}
\arguments{
\item{no_value}{}
}
\description{
Title
}
