% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{db_connect_to_access}
\alias{db_connect_to_access}
\title{Title}
\usage{
db_connect_to_access(
  access_file,
  uid = "",
  pwd = "",
  encoding = "",
  package = c("DBI", "RODBC"),
  check_connection = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
