% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{db_connect_to_access_string}
\alias{db_connect_to_access_string}
\title{Title}
\usage{
db_connect_to_access_string(access_file, uid = "", pwd = "", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
