% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_database_archive_functions.R
\name{db_copy_new_files}
\alias{db_copy_new_files}
\title{Title}
\usage{
db_copy_new_files(
  input_dir,
  db_dir_base,
  k_id,
  v_id,
  seq_start,
  overwrite = FALSE,
  execute = TRUE,
  sep = ".",
  file_extension = ".jpg"
)
}
\arguments{
\item{file_extension}{}
}
\description{
Title
}
