% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{db_import_access}
\alias{db_import_access}
\title{Title}
\usage{
db_import_access(
  con,
  sql_expr,
  ...,
  package = c("DBI", "RODBC"),
  tibble = TRUE,
  na_value = c("", " ", "  ")
)
}
\arguments{
\item{na_value}{}
}
\description{
Title
}
