% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{db_import_access_n}
\alias{db_import_access_n}
\title{Title}
\usage{
db_import_access_n(
  fls,
  sql_expr,
  ...,
  source_column = "source",
  source_names = basename(fls),
  na_value = c("", " ", "  ")
)
}
\arguments{
\item{na_value}{}
}
\description{
Title
}
