% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_database_archive_functions.R
\name{db_insert_archive}
\alias{db_insert_archive}
\title{Title}
\usage{
db_insert_archive(
  k_id,
  v_id,
  seq_start = 1,
  observer_id,
  observer_text,
  data_type_id,
  document_type_id,
  time_text,
  dsn = "Havsorn_Data",
  db_dir_base,
  pattern = "*.jpg",
  execute = FALSE
)
}
\arguments{
\item{execute}{}
}
\description{
Title
}
