% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{db_pquery}
\alias{db_pquery}
\title{Title}
\usage{
db_pquery(
  tbl,
  flds,
  type = c("insert", "select", "update", "delete"),
  where_fld = NULL
)
}
\arguments{
\item{where_fld}{}
}
\description{
Title
}
