% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helper_functions.R
\name{download_zip}
\alias{download_zip}
\title{Title}
\usage{
download_zip(
  wd,
  folder,
  files,
  files_url,
  file_type = "zip",
  files_unzip = TRUE,
  delete_zip = TRUE,
  junkpaths = TRUE,
  overwrite = TRUE,
  add_date = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
