% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_handling.R
\name{exiftool_read}
\alias{exiftool_read}
\title{Title}
\usage{
exiftool_read(
  path,
  args = c("-j", "-q", "-n", "-alldates"),
  charset = "cp1252",
  tags = c("filename", "filesize", "imagesize"),
  exiftool_path = "exiftool(-k).exe",
  args_file = "tmp_exif_data.arg",
  output_file = "tmp_exif_data.json",
  convert_encoding = TRUE,
  current_encoding = "utf-8",
  target_encoding = "iso-8859-1",
  clean_names = TRUE,
  delete_temp_files = TRUE
)
}
\arguments{
\item{delete_temp_files}{}
}
\description{
Title
}
