% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_models_gam.R
\name{geom_trend_gam}
\alias{geom_trend_gam}
\title{Title}
\usage{
geom_trend_gam(
  ...,
  rect.params = list(),
  fit.params = list(),
  ci.params = list(),
  fit.decr.params = list(),
  fit.incr.params = list(),
  hline.params = list(),
  point.params = list(),
  add_items = c("ci", "fit", "rect", "hline", "fit.decr", "fit.incr", "point"),
  object,
  n_pts_predict = 300,
  term_index = 1
)
}
\arguments{
\item{term_index}{}
}
\description{
Title
}
