% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_database_archive_functions.R
\name{get_seq_start}
\alias{get_seq_start}
\title{Title}
\usage{
get_seq_start(
  db_dir_base,
  db_dir_path,
  pattern = "*.jpg",
  start = 12,
  end = 15
)
}
\arguments{
\item{end}{}
}
\description{
Title
}
