% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_models_gam.R
\name{plot_gam_fit_trends}
\alias{plot_gam_fit_trends}
\title{Title}
\usage{
plot_gam_fit_trends(
  object,
  x,
  y,
  group,
  p_value_cutoff,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{ylab}{}
}
\description{
Title
}
