% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_fagel3_helper_functions.R
\name{rc_coords}
\alias{rc_coords}
\title{Title}
\usage{
rc_coords(
  latitude,
  longitude,
  outformat = c("dm", "dms"),
  pad_output = TRUE,
  width = 6,
  pad = " ",
  side = "right",
  sep = " "
)
}
\arguments{
\item{sep}{}
}
\description{
Title
}
