% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_fagel3_search_txt.R
\name{rc_data_search_txt}
\alias{rc_data_search_txt}
\title{Title}
\usage{
rc_data_search_txt(
  pattern,
  path_pattern = ".*RCdata\\\\.txt$",
  fixed = FALSE,
  markers = FALSE
)
}
\arguments{
\item{markers}{}
}
\description{
Title
}
