% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_fagel3_helper_functions.R
\name{rc_fagel3_report_file}
\alias{rc_fagel3_report_file}
\title{Title}
\usage{
rc_fagel3_report_file(
  mnr = 658,
  year_filter = 2022,
  .markare = markare,
  .medhjalpare = medhjalpare,
  .signaturer = signaturer,
  .lokaler = lokaler,
  .kontr = kontr,
  .kullar = kullar,
  .ringon = ringon,
  .ringar = ringar,
  export = TRUE
)
}
\arguments{
\item{export}{}
}
\description{
Title
}
