% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_fagel3_helper_functions.R
\name{rc_import_from_db}
\alias{rc_import_from_db}
\title{Title}
\usage{
rc_import_from_db(
  access_file,
  sql_expr,
  clean_spaces = NULL,
  clean_names = TRUE,
  date_transform = TRUE,
  package = c("DBI", "RODBC"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
