% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_fagel3_helper_functions.R
\name{rc_parse_province_midpt}
\alias{rc_parse_province_midpt}
\title{Title}
\usage{
rc_parse_province_midpt(
  .x,
  pattern =
    "(?<latd>[0-9]{2})(?<latm>[0-9]{2})(?<lath>[A-Z]{1})(?<lond>[0-9 ]+)(?<lonm>[0-9]{2})(?<lonh>[A-Z]{1})",
  col_names = c("lat_dd", "lon_dd")
)
}
\arguments{
\item{col_names}{}
}
\description{
Title
}
