% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_fagel3_helper_functions.R
\name{rc_parse_ring}
\alias{rc_parse_ring}
\title{Title}
\usage{
rc_parse_ring(.x, pattern = "[a-zA-Z]+", width = 8, side = "left", pad = " ")
}
\arguments{
\item{pad}{}
}
\description{
Title
}
