% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_fagel3_search_txt.R
\name{rc_ringon_search_txt}
\alias{rc_ringon_search_txt}
\title{Title}
\usage{
rc_ringon_search_txt(
  species_code,
  year = NULL,
  pattern_sites = "^L\\\\|",
  fixed = FALSE,
  markers = FALSE,
  empty_ringondb = "W:/projects/data/ringing/RingDb_tom_2018-10-15.mdb"
)
}
\arguments{
\item{empty_ringondb}{}
}
\description{
Title
}
