% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helper_functions.R
\name{round_up}
\alias{round_up}
\title{Round a numeric value upwards}
\usage{
round_up(.x)
}
\arguments{
\item{.x}{Numeric value}
}
\value{
Numeric
}
\description{
This functions rounds a numeric value upwards
}
