% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_database_archive_functions.R
\name{set_file_name}
\alias{set_file_name}
\title{Title}
\usage{
set_file_name(k_id, v_id, b_id, sep = ".", file_extension = ".jpg")
}
\arguments{
\item{file_extension}{}
}
\description{
Title
}
