% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_concatenate_df_and_strings.R
\name{str_c_unique}
\alias{str_c_unique}
\title{Title}
\usage{
str_c_unique(.x, collapse = "; ", sort = FALSE, unique = TRUE, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\description{
Title
}
