% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_database_archive_functions.R
\name{thumbs_bat}
\alias{thumbs_bat}
\title{Title}
\usage{
thumbs_bat(
  files,
  output_file,
  output_dir = getwd(),
  sub_dir = TRUE,
  relative = FALSE,
  file_encoding = "latin1",
  i_view_path = "C:/Program Files/IrfanView/i_view64.exe"
)
}
\arguments{
\item{i_view_path}{}
}
\description{
Title
}
