% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_indicator.R
\name{wtse_boot}
\alias{wtse_boot}
\title{Title}
\usage{
wtse_boot(
  .x,
  outcomes = c("0C", "1", "2", "3", "0G", ">0", ">=1", ">=2", ">=3"),
  n_boot = 1000
)
}
\arguments{
\item{n_boot}{}
}
\description{
Title
}
