% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_summarize_reproduction_from_db.R
\name{wtse_columns}
\alias{wtse_columns}
\title{Title}
\usage{
wtse_columns(
  dsn = "Havsorn",
  table = "luOvervakningUtfall",
  column = "OvervakningUtfallR",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
