% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_gis_join_areas.R
\name{wtse_join_areas}
\alias{wtse_join_areas}
\title{Title}
\usage{
wtse_join_areas(
  .x,
  .y,
  .x_group = LokalID,
  .y_ID,
  .y_NAME,
  OmradeTypID,
  MetodAvstand = 0,
  Metod = st_intersects,
  MetodAndelArea = 1,
  left = FALSE,
  geometry_field = geom
)
}
\arguments{
\item{geometry_field}{}
}
\description{
Title
}
