% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_indicator.R
\name{wtse_productivity}
\alias{wtse_productivity}
\title{Title}
\usage{
wtse_productivity(
  dd,
  stats.brood = TRUE,
  bootstrap.brood = TRUE,
  n.boot = 10000,
  replace = TRUE
)
}
\arguments{
\item{replace}{}
}
\description{
Title
}
