% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_gis_sites.R
\name{wtse_sites}
\alias{wtse_sites}
\title{Title}
\usage{
wtse_sites(
  spatial = FALSE,
  crs = NULL,
  add_coordinates = FALSE,
  coordinate_cols = c("x", "y"),
  odbc_name = "Havsorn_Data",
  encoding = "",
  add_subsites = TRUE,
  add_monitoring = TRUE,
  drop_id_columns = TRUE,
  ortnamn_path = "G:/Ortnamn/GSD-Ortnamn_2012.gpkg"
)
}
\arguments{
\item{ortnamn_path}{}
}
\description{
Title
}
