% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_database_access_sites.R
\name{wtse_sites_get_access}
\alias{wtse_sites_get_access}
\title{Title}
\usage{
wtse_sites_get_access(
  area_field,
  area_id,
  kontakt_id,
  wtse_sites = sites,
  dsn = "Havsorn_Data"
)
}
\arguments{
\item{dsn}{}
}
\description{
Title
}
