% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_database_access_sites.R
\name{wtse_sites_get_access_map}
\alias{wtse_sites_get_access_map}
\title{Title}
\usage{
wtse_sites_get_access_map(.data, wtse_sites = sites_sf)
}
\arguments{
\item{wtse_sites}{}
}
\description{
Title
}
