% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_gis_sites_grid.R
\name{wtse_sites_grid}
\alias{wtse_sites_grid}
\title{Title}
\usage{
wtse_sites_grid(
  .x,
  .grid,
  type = c("LINESTRING", "POINT"),
  position = c("centroid", "lower-left")
)
}
\arguments{
\item{position}{}
}
\description{
Title
}
