% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_summarize_reproduction_from_db.R
\name{wtse_transform}
\alias{wtse_transform}
\title{Title}
\usage{
wtse_transform(.data, ..., .response = UtfallR, .columns = NULL, .drop = FALSE)
}
\arguments{
\item{.drop}{}
}
\description{
Title
}
