% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtse_stats_trend.R
\name{wtse_trend}
\alias{wtse_trend}
\title{Title}
\usage{
wtse_trend(
  y,
  x,
  model,
  reference = c(0, 0, 0),
  npts = 500,
  alpha = 0.05,
  pch = 21,
  pch_col = "black",
  pch_bg = "steelblue",
  pch_cex = 1,
  add = FALSE,
  add_points = TRUE,
  add_line = TRUE,
  ref_line = TRUE,
  ref_col = "darkgrey",
  ref_lty = c(3, 2, 3),
  ref_lwd = c(1.5, 1.5, 1.5),
  ref_polygon = FALSE,
  ref_polygon.col = "yellow",
  trend_col = "black",
  trend_lty = 1,
  trend_lwd = 1.5,
  signif = TRUE,
  signif_lwd = 2,
  signif_col = c("red", "blue"),
  ci = TRUE,
  main = "White-tailed sea eagle",
  xlab = "Year",
  ylab = "Variable",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
